# pet-booking-be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservation API",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get paginated list of reservations",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Reservation"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total_pages": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "total_items": {
                                                            "type": "integer",
                                                            "example": 50
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "description": "Creates a new reservation with the provided details.",
                "operationId": "storeReservation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "property": null
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get a reservation by ID",
                "description": "Returns a reservation record if the ID exists, otherwise returns null.",
                "operationId": "getReservationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation found or not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "description": "Reservation object or null",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Reservation"
                                                },
                                                {
                                                    "type": "null",
                                                    "example": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation by ID",
                "description": "Deletes a reservation by its ID. Returns success status.",
                "operationId": "deleteReservationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation deleted successfully."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation not found."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ReservationCreateRequest": {
                "required": [
                    "frequency",
                    "start_date",
                    "days",
                    "times"
                ],
                "properties": {
                    "frequency": {
                        "type": "string",
                        "enum": [
                            "recurring",
                            "one-time"
                        ],
                        "example": "recurring"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-06"
                    },
                    "days": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "mon",
                                "tue",
                                "wed",
                                "thu",
                                "fri",
                                "sat",
                                "sun"
                            ]
                        },
                        "example": [
                            "mon",
                            "tue"
                        ]
                    },
                    "times": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "morning",
                                "afternoon",
                                "evening"
                            ]
                        },
                        "example": [
                            "morning",
                            "afternoon"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 300,
                        "example": "Just have fun"
                    }
                },
                "type": "object"
            },
            "Reservation": {
                "required": [
                    "frequency",
                    "start_date",
                    "days",
                    "times"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "frequency": {
                        "type": "string",
                        "example": "recurring"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-06"
                    },
                    "days": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "mon"
                        },
                        "example": [
                            "mon",
                            "tue"
                        ]
                    },
                    "times": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "morning"
                        },
                        "example": [
                            "morning",
                            "afternoon",
                            "evening"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Just have fun"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Reservations",
            "description": "Reservations"
        }
    ]
}